refactor(validator): move axon-fallback helpers to chronoseek/utils.py (#32)
resolve_axon_endpoint() and build_endpoint_map_with_axon_fallback() move
out of chronoseek/chutes/runtime.py into a new chronoseek/utils.py, since
axon-vs-Chutes endpoint selection isn't Chutes-specific logic. runtime.py
keeps only the Chutes-commitment resolution (resolve_submission_endpoint,
build_submission_endpoint_map, build_evaluation_endpoints) and the health
check/query helpers unchanged.
